requirements.txt
anthropic>=0.97.0
python-dotenv>=1.0.0
flask>=1.1.0
